requirements.txt
streamlit
pytorch
numpy
Pillow
torchvision